requirements.txt
fastapi
uvicorn
google-generativeai
pillow
volcengine-python-sdk[ark]
pydantic
python-multipart
